Practice6/requirements.txt
pdfminer.six
python-docx
tqdm